+++ include/sys/process.h
@@ -158,7 +158,7 @@ static inline pid_t process_self(void)
 static inline int process_selfname(char* name, size_t size)
 {
 #if defined(OS_WINDOWS)
-	if(0 == GetModuleFileNameA(NULL, name, size))
+	if(0 == GetModuleFileNameA(NULL, name, (DWORD)size))
 		return (int)GetLastError();
 	return 0;
 #else
@@ -178,7 +178,7 @@ static inline int process_name(pid_t pid, char* name, size_t size)
 	if(!h)
 		return (int)GetLastError();
 
-	r = GetModuleFileNameExA(h, NULL, name, size);
+	r = GetModuleFileNameExA(h, NULL, name, (DWORD)size);
 	CloseHandle(h);
 	return 0==r ? (int)GetLastError() : 0;
 #else
